Bot_data_base/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.7.0
openai==1.30.0
chromadb==0.5.0
sentence-transformers==3.0.0
tiktoken==0.7.0
yt-dlp==2024.5.27
pyyaml==6.0.1
python-dotenv==1.0.1
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
jinja2==3.1.4
aiofiles==23.2.1
numpy==1.26.4
